Added function d and function e
d-Find the average of the array. Display how each element of the array differs from the average. For example, if the array is {2, 6, 9, 3, 5}, and the average is 5, then the differences from the average would be {-3, 1, 4, -2, 0}.
e-Find the sum of elements with odd- and even-numbered indexes. For example, if the array is {1, 5, 8, 9, 10, 2, 6}, the sum of the odd indexes is 5 + 9 + 2, and the sum of the even indexes is 1 + 8 + 10 + 6, assuming indexes start at 0.